refactor: OrderProduct.price를 Long에서 Integer로 변경

+++ apps/commerce-api/src/main/java/com/loopers/domain/order/OrderCommand.java
@@ -21,7 +21,7 @@ public final class OrderCommand {
         public static class Product {
             private final Long productOptionId;
             private final Integer quantity;
-            private final Long price;
+            private final Integer price;
         }
     }
